# Conflicts:
#	knitkit/app/controllers/knitkit/erp_app/desktop/website_controller.rb
#	knitkit/app/models/website.rb
#	knitkit/knitkit.gemspec

+++ erp_tech_svcs/lib/erp_tech_svcs/file_support/s3_manager.rb
@@ -141,10 +141,19 @@ module ErpTechSvcs
         path = path.sub(%r{^/}, '')
         result = false
         message = nil
+
         begin
-          bucket.objects.with_prefix(path).delete_all
-          message = "File was deleted successfully"
-          result = true
+          if is_directory && (self.build_tree(path)[:children].count != 0)
+            message = ErpTechSvcs::FileSupport::Manager::FOLDER_IS_NOT_EMPTY
+            result = false
+
+          else
+            bucket.objects.with_prefix(path).delete_all
+            message = "File was deleted successfully"
+            result = true
+
+          end
+
         rescue => ex
           result = false
           message = ex
